+++ include/internal/noscrypt_logger.hpp
